Began Integration of Physics System
Now that the math primitives have been created, I began the process of
integrating the physics system into Grand Central Dispatch. Not done yet,
taking a break for now.

+++ src/rendersystem.cpp
@@ -30,7 +30,7 @@ void RenderSystem::addNode(int key, Mesh data)
     nodes.push_back(newNode);
 }
 
-void RenderSystem::tick(__attribute__((unused)) float dt)
+void RenderSystem::tick(__attribute__((unused)) double dt)
 {
     // Clear the frame buffer, depth buffer, and stencil buffer
     glClearColor(0.0, 0.0, 0.0, 1.0);
